Ignore incomplete escape sequence
After starting an external tool an escape sequence for e.g.
a cursor key may not be recognized correctly. Parts of this
sequence may be recognized as commands, e.g. copy command
"C" from sequence "^[[C...". If such a case is detected, all
keys are ignored until ENTER is pressed.